refactor: consolidate runtime helpers and adapter normalization
