+++ .changeset/source-polish-backlog.md
@@ -1,0 +1,8 @@
+---
+'@openai/agents-core': patch
+'@openai/agents-openai': patch
+'@openai/agents-realtime': patch
+'@openai/agents-extensions': patch
+---
+
+refactor: consolidate internal runtime helpers and adapter normalization
