+++ packages/agents-core/src/handoff.ts
@@ -210,7 +210,7 @@ export class Handoff<
    * sees the entire conversation history. In some cases, you may want to filter inputs e.g. to
    * remove older inputs, or remove tools from existing inputs.
    *
-   * The function will receive the entire conversation hisstory so far, including the input item
+   * The function will receive the entire conversation history so far, including the input item
    * that triggered the handoff and a tool call output item representing the handoff tool's output.
    *
    * You are free to modify the input history or new items as you see fit. The next agent that runs
