+++ packages/agents-core/src/sandbox/shared/typeGuards.ts
@@ -12,10 +12,10 @@ export function isStringRecord(
 }
 
 export function readOptionalString(
-  value: Record<string, unknown>,
+  value: Record<string, unknown> | undefined,
   key: string,
 ): string | undefined {
-  const result = value[key];
+  const result = value?.[key];
   return typeof result === 'string' ? result : undefined;
 }
 
@@ -38,4 +38,38 @@ export function readStringArray(value: unknown): string[] {
   return Array.isArray(value)
     ? value.filter((entry): entry is string => typeof entry === 'string')
     : [];
+}
+
+export function readOptionalRecord(
+  value: unknown,
+): Record<string, unknown> | undefined {
+  return isRecord(value) ? value : undefined;
+}
+
+export function readOptionalNumber(
+  value: Record<string, unknown>,
+  key: string,
+): number | undefined {
+  const result = value[key];
+  return typeof result === 'number' ? result : undefined;
+}
+
+export function readOptionalBoolean(
+  value: Record<string, unknown>,
+  key: string,
+): boolean | undefined {
+  const result = value[key];
+  return typeof result === 'boolean' ? result : undefined;
+}
+
+export function readOptionalRecordArray(
+  value: unknown,
+): Array<Record<string, unknown>> | undefined {
+  return Array.isArray(value) ? value.filter(isRecord) : undefined;
+}
+
+export function readOptionalStringRecord(
+  value: unknown,
+): Record<string, string> | undefined {
+  return isStringRecord(value) ? value : undefined;
 }
