+++ packages/agents-core/test/utils/index.test.ts
@@ -1,8 +1,18 @@
 import { describe, it, expect } from 'vitest';
-import { toSmartString } from '../../src/utils/index';
+import * as Utils from '../../src/utils/index';
+import { formatInlineData, getInlineMediaType } from '../../src/utils/internal';
 
 describe('utils/index', () => {
   it('toSmartString', () => {
-    expect(toSmartString('foo')).toBe('foo');
+    expect(Utils.toSmartString('foo')).toBe('foo');
+  });
+
+  it('keeps inline-data helpers on the internal surface', () => {
+    expect('formatInlineData' in Utils).toBe(false);
+    expect('getInlineMediaType' in Utils).toBe(false);
+    expect(formatInlineData('YWJj', 'image/png')).toBe(
+      'data:image/png;base64,YWJj',
+    );
+    expect(getInlineMediaType({ mimeType: 'image/png' })).toBe('image/png');
   });
 });
