+++ packages/agents-openai/src/responsesWebSocketConnection.ts
@@ -22,16 +22,25 @@ export type ResponsesWebSocketKeepAliveOptions = {
   pingTimeoutMs?: number | null;
 };
 
+export const RESPONSES_WEBSOCKET_ERROR_MESSAGES = {
+  connection_closed_before_opening:
+    'Responses websocket connection closed before opening.',
+  connection_closed_before_terminal_response_event:
+    'Responses websocket connection closed before a terminal response event.',
+  pong_timeout: 'Responses websocket pong timeout.',
+  socket_not_open: 'Responses websocket is not open.',
+} as const;
+
 export type ResponsesWebSocketInternalErrorCode =
-  | 'connection_closed_before_opening'
-  | 'connection_closed_before_terminal_response_event'
-  | 'pong_timeout'
-  | 'socket_not_open';
+  keyof typeof RESPONSES_WEBSOCKET_ERROR_MESSAGES;
 
 export class ResponsesWebSocketInternalError extends Error {
   readonly code: ResponsesWebSocketInternalErrorCode;
 
-  constructor(code: ResponsesWebSocketInternalErrorCode, message: string) {
+  constructor(
+    code: ResponsesWebSocketInternalErrorCode,
+    message: string = RESPONSES_WEBSOCKET_ERROR_MESSAGES[code],
+  ) {
     super(message);
     this.name = 'ResponsesWebSocketInternalError';
     this.code = code;
@@ -141,9 +150,10 @@ export function shouldWrapNoEventWebSocketError(error: unknown): boolean {
   }
 
   return (
-    error.message === 'Responses websocket connection closed before opening.' ||
     error.message ===
-      'Responses websocket connection closed before a terminal response event.'
+      RESPONSES_WEBSOCKET_ERROR_MESSAGES.connection_closed_before_opening ||
+    error.message ===
+      RESPONSES_WEBSOCKET_ERROR_MESSAGES.connection_closed_before_terminal_response_event
   );
 }
 
@@ -156,7 +166,7 @@ export function isWebSocketNotOpenError(error: unknown): boolean {
     return false;
   }
 
-  if (error.message === 'Responses websocket is not open.') {
+  if (error.message === RESPONSES_WEBSOCKET_ERROR_MESSAGES.socket_not_open) {
     return true;
   }
 
@@ -310,7 +320,6 @@ export class ResponsesWebSocketConnection {
     ) {
       throw new ResponsesWebSocketInternalError(
         'connection_closed_before_opening',
-        'Responses websocket connection closed before opening.',
       );
     }
 
@@ -334,7 +343,6 @@ export class ResponsesWebSocketConnection {
           this.#error ??
             new ResponsesWebSocketInternalError(
               'connection_closed_before_opening',
-              'Responses websocket connection closed before opening.',
             ),
         );
       };
@@ -363,10 +371,7 @@ export class ResponsesWebSocketConnection {
 
   async send(data: string): Promise<void> {
     if (this.#socket.readyState !== this.#socket.OPEN) {
-      throw new ResponsesWebSocketInternalError(
-        'socket_not_open',
-        'Responses websocket is not open.',
-      );
+      throw new ResponsesWebSocketInternalError('socket_not_open');
     }
 
     this.#socket.send(data);
@@ -486,10 +491,7 @@ export class ResponsesWebSocketConnection {
       }
       this.#pingTimeoutTimer = setTimeout(() => {
         this.#pingTimeoutTimer = undefined;
-        this.#error = new ResponsesWebSocketInternalError(
-          'pong_timeout',
-          'Responses websocket pong timeout.',
-        );
+        this.#error = new ResponsesWebSocketInternalError('pong_timeout');
         try {
           if (typeof socket.terminate === 'function') {
             socket.terminate();
